Status components moved to @/components/status/
