+++ assets/js/pages/ProjectPage/CheckInSection.tsx
@@ -10,7 +10,7 @@ import { DimmedLabel } from "./Label";
 import { SecondaryButton } from "@/components/Buttons";
 import { Link } from "@/components/Link";
 import { Paths } from "@/routes/paths";
-import { SmallStatusIndicator } from "@/features/projectCheckIns/SmallStatusIndicator";
+import { SmallStatusIndicator } from "@/components/status";
 import { Summary } from "@/components/RichContent";
 
 export function CheckInSection({ project }: { project: Projects.Project }) {
